Rename rpc_address to rpc_socket
