+++ mullvad-paths/src/lib.rs
@@ -51,8 +51,8 @@ pub use logs::{get_log_dir, log_dir};
 pub mod resources;
 pub use resources::get_resource_dir;
 
-mod rpc_address;
-pub use rpc_address::get_rpc_socket_path;
+mod rpc_socket;
+pub use rpc_socket::get_rpc_socket_path;
 
 mod settings;
 pub use settings::settings_dir;
